Restructure project
Created a top-level project to contain sample modules.
This lets us create dependencies between the different
samples, and hopefully keeps things nice and tidy where
Android Studio is concerned.

The downside to doing it this way is that all of the samples
get loaded into one big project, so if we end up with a truly
large number of samples we may want to rethink this.