Add state and parameter management
Removed global variables from HelloLowLatencyOut by
providing thread-safe (I think!) access to two blocks of
memory: one for processing state, and one for parameters.

Access to the state block is only granted via callbacks, so the
library can prevent simultaneous access (as long as nobody
does anything stupid like store off the pointer for later).

The parameter block is implemented by a kind of
brain-dead variant on Glenn's SingleStateQueue. It pushes
responsibility for retries out to the caller, so the core
implementation can CAS and bail. Aside from that, it
makes the convenient assumption that parameters are
always updated as a block. This seems like a reasonable
simplification for an API that assumes the parameters are
defined by the same code that does the processing.
Oh, and it's also in C++11, because of course it is.

This CL also adds the optimistically named project
"HelloLowLatencyIO," but it doesn't yet contain actual
full-duplex functionality.