Added full duplex processing
"HelloLowLatencyIO" now implements what you might call
the smallest possible physical modeling synth: an adjustable-
cutoff LPF set to a really high Q. It filters the input and
feeds it to the output.

There's still lots of ugliness in this sample, especially
magic numbers and an assumption that samples are shorts.
I'll have those fixed shortly.

Change-Id: Ic86895379d0046accc051a225cf6433743a51c57