Improved parameter delivery and added device characteristics accessor
Added a C function to get the device characteristics, at
@gfan's request.

Rewrote the parameter pipe (again) to use a set of mutexes. After
discussion with C++ experts it's clear that we can't realize the original
seqlock-based design in standard C++. Instead, both reader and writer
acquire locks, but the reader does not block. It uses try_lock, which is
implemented by a single CAS, and therefore cannot block.

Change-Id: Ifb13ffcb6a061e9ce4ffcc9d592937e315a41ac3