Moved OpenSL calls to a background thread
This moves expensive OpenSL calls (which, by the looks of it,
is all of them) to a single background thread via a FIFO. This keeps
the call order serial while preventing OpenSL from blocking the
application thread.

One consequence of this change is that the OnDeviceChanged callback
will now be called on the worker thread rather than the application
thread.

Change-Id: If4805b89ffd7ef4acebda698a102b3ded9547b49